RavenDB-20417 & RavenDB-20431 address PR comments
